feat: add poste field to Lead forms and model.
This commit includes the addition of the 'poste' field across the Lead form, model and related translations. The updates have been made in the form template, form type, model classes, controller and messages.fr.yml file for French translations. The form field is not required and allows users to specify their current position or job title.